requirements.txt
nose>=1.1.2
scikit-learn>=0.17
pillow>=5.1.0
numpy>=1.8.0
scipy>=1.0.1
